Addedd structure  to manage differents types of identifier. Still to do the json composing for L3 protocol (many ip for one interface)
