+++ nodeshot/core/base/utils.py
@@ -116,49 +116,3 @@ def after(date, **kwargs):
     :paramm datetime: datetime object to which add more time
     """
     return date + timedelta(**kwargs)
-
-
-def update_topology():
-    for topology in Topology.objects.all():
-        module = importlib.import_module(topology.backend)
-        parser_class_name = topology.backend.split('.')[-1]
-        classparsr = getattr(module, parser_class_name)
-        njparser = NetJsonParser(to_netjson(topology))
-        parser = classparser(topology.url)
-        graph_diff = diff(njparser, parser)
-        for link in graph_diff['added']:
-            if_a = Interface.objects.get(mac=link[0])
-            if_b = Interface.objects.get(mac=link[1])
-            Link.objects.create(interface_a=if_a, interface_b=if_b, metric_value=link[2]['weight'])
-        for link in graph_diff['removed']:
-            if_a = Interface.objects.get(mac=link[0])
-            if_b = Interface.objects.get(mac=link[1])
-            try:
-                l = Link.objects.get(interface_a=if_a, interface_b=if_b)
-                l.delete()
-            except Link.DoesNotExist:
-                pass
-
-
-def exist_node(id, nodes):
-    return any([node.get('id') == id for node in nodes])
-
-
-def to_netjson(topology):
-    nodes = []
-    links = []
-
-    for link in Link.objects.filter(topology=topology):
-        node = {'id': link.inteface_a.mac}
-        if not exist_node(node, nodes):
-            nodes.append(node)
-        links.append(link.to_netjson())
-
-    NetJson = {'type': 'NetworkGraph',
-                'protocol': island.get_protocol_display(),
-                'version': '0',
-                'metric': '0',
-                'router_id': island.url,
-                'nodes': nodes,
-                'links': links}
-    return NetJson
