+++ nodeshot/networking/links/models/choices.py
@@ -22,3 +22,16 @@ METRIC_TYPES = {
     'ETC': 'etc',
     'HOP': 'hop'
 }
+
+ID_TYPES = {
+    'IPv4/v6': 1,
+    'MAC': 2,
+}
+
+NETDIFF_PARSERS = getattr(settings, 'NODESHOT_NETDIFF_PARSERS', [
+    ('netdiff.OlsrParser', 'Olsr (jsoninfo)'),
+    ('netdiff.BatmanParser', 'Batman (Alfred)'),
+    ('netdiff.BmxParser', 'Bmx6 (q6m)'),
+    ('netdiff.NetJsonParser', 'Netjson (network-graph)'),
+    ('netdiff.CnmlParser', 'CNML'),
+])
